% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tools.R
\name{get_leaf_node}
\alias{get_leaf_node}
\title{Find the leaf node for a test sample.}
\usage{
get_leaf_node(tree, newdata, node.id = TRUE)
}
\arguments{
\item{tree}{A LASERF tree object (retrieved by \code{get_tree}).}

\item{newdata}{Points at which leaf predictions should be made.}

\item{node.id}{Boolean indicating whether to return the node.id for each query sample (default), or
if FALSE, a list of node numbers with the samples contained.}
}
\value{
A vector of integers indicating the leaf number for each sample in the given tree.
}
\description{
Given a LASERF tree object, compute the leaf node a test sample falls into. The nodes in a LASERF tree
are numbered breadth first, and the returned numbers will be the leaf integer according
to this ordering. To get kernel weights based on leaf membership, see the function
\code{\link{get_forest_weights}}.
}
\examples{
\donttest{
# TODO
}

}
