% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laserf-package.R
\docType{package}
\name{laserf-package}
\alias{laserf-package}
\title{laserf: Locally Adaptive Subspace Estimation (with) Random Forests}
\description{
A description of laserf TODO
}
\examples{
\donttest{
  # TODO
}
}
\author{
\strong{Maintainer}: David Fleischer \email{david.p.fleischer@gmail.com}

}
\keyword{internal}
