% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laserf_options.R
\name{laserf_options}
\alias{laserf_options}
\title{laserf package options}
\usage{
laserf_options()
}
\value{
Prints the current laserf package options.
}
\description{
laserf package options can be set using R's \code{\link{options}} command.
The current available options are:
\itemize{
\item \code{laserf.legacy.seed}: controls whether laserf's random seed behavior depends on
the number of CPU threads used to train the forest. The default value is \code{FALSE}.
Set to \code{TRUE} to recover random seed behaviour inherited from grf versions prior to 2.4.0.
}
}
\examples{
\donttest{
# Use random seed behavior prior to version 2.4.0.
options(laserf.legacy.seed = TRUE)

# Print current package options.
laserf_options()

# Use random seed independent of num.threads (default).
options(laserf.legacy.seed = FALSE)
}

}
