% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laserf.R
\name{make_laserf_preds}
\alias{make_laserf_preds}
\title{Form the subspace forest predictions given the raw predictions returned by C++}
\usage{
make_laserf_preds(
  preds.raw,
  num.features,
  rank,
  Y = NULL,
  compute.scores = FALSE,
  compute.projections = FALSE
)
}
\arguments{
\item{preds.raw}{TODO...}

\item{num.features}{TODO... dimension of Y input feature space}

\item{rank}{TODO... rank}

\item{Y}{TODO... Default \code{NULL}}

\item{compute.scores}{TODO... low-dimensional representations... Ignored if \code{compute.projections = TRUE}}

\item{compute.projections}{TODO... projections}
}
\value{
TODO...
}
\description{
Parse and partition the raw prediction vectors returned by C++. Reform the local
eigenvector matrix (principal components), alongside separate data structures
for the local eigenvalues (variances) and the local feature-space means.
}
\details{
When new testing features \code{Y} are supplied, can also compute the predicted score
and projection vectors given \code{Y} at the fitted local eigenvectors and means.
}
\keyword{internal}
