% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laserf.R
\name{predict.laserf}
\alias{predict.laserf}
\title{Predict with a subspace forest}
\usage{
\method{predict}{laserf}(
  object,
  newX = NULL,
  newY = NULL,
  rank = NULL,
  compute.scores = FALSE,
  compute.projections = FALSE,
  num.threads = NULL,
  ...
)
}
\arguments{
\item{object}{The trained forest.}

\item{newX}{TODO... C++ just takes the test auxiliary covariates X, but in R we can accept a test Y as well}

\item{newY}{TODO... See \code{newX}}

\item{rank}{TODO... Optional, defaults to the trained \code{split.rank}}

\item{compute.scores}{TODO... compute z = Vt y}

\item{compute.projections}{TODO... compute yhat = Vz}

\item{num.threads}{Number of threads used in prediction. If set to NULL, the software
automatically selects an appropriate amount.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A list containing \code{predictions}: a matrix of predictions for each outcome.
}
\description{
TODO
}
\examples{
\donttest{
# TODO ...
}

}
