% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.laserf_tree}
\alias{print.laserf_tree}
\title{Print a LASERF tree object.}
\usage{
\method{print}{laserf_tree}(x, ...)
}
\arguments{
\item{x}{The tree to print.}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Print a LASERF tree object.
}
