% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laserf.R
\name{restructure_subspace_predictions}
\alias{restructure_subspace_predictions}
\title{Restructure the vectorized subspace prediction vectors}
\usage{
restructure_subspace_predictions(
  preds.raw,
  num.features,
  rank,
  Y = NULL,
  compute.scores = FALSE,
  compute.projections = FALSE
)
}
\arguments{
\item{preds.raw}{TODO...}

\item{num.features}{TODO... dimension of Y input feature space}

\item{rank}{TODO... rank}

\item{Y}{TODO...}

\item{compute.scores}{TODO... low-dimensional representations... Ignored if \code{compute.projections = TRUE}}

\item{compute.projections}{TODO... projections}
}
\value{
TODO...
}
\description{
Restructure the vectorized subspace prediction vectors
}
\keyword{internal}
